refactor: 지갑 페이지 컴포넌트들 분리
