+++ src/pages/Wallet.jsx
@@ -1,6 +1,7 @@
 import React, { useContext } from 'react';
 
-import WalletItem from '../components/WalletItem';
+import WalletItemList from '../components/WalletItemList';
+import TotalMoney from '../components/TotalMoney';
 import { MoneyContext } from '../context/MoneyProvider';
 import { PriceContext } from '../context/PriceProvider';
 
@@ -21,39 +22,10 @@ export default function Wallet() {
     setRemainMoney(remainMoney + won);
   };
 
-  const addUnitNotation = (element, index, array) => {
-    const unitPosition = 3;
-    const isHighestDigit = index + 1 === array.length;
-    const hasUnitNotation = (index + 1) % unitPosition === 0 && !isHighestDigit;
-
-    return hasUnitNotation ? `,${element}` : element;
-  };
-
-  const toatlPrice = moneyState
-    .map(({ won, num }) => won * num)
-    .reduce((aMoney, bMoney) => aMoney + bMoney);
-
-  const totalPriveWithUnitNotation = `${toatlPrice}`
-    .split('')
-    .reverse()
-    .map((element, index, array) => addUnitNotation(element, index, array))
-    .reverse()
-    .join('');
-
   return (
     <>
-      <ul>
-        {moneyState.map(({ won, num }) => (
-          <WalletItem
-            key={`money-${won}`}
-            icon={`${won}원 `}
-            won={won}
-            num={num}
-            onClick={() => handleClickWon(won, num)}
-          />
-        ))}
-      </ul>
-      <span>{`${totalPriveWithUnitNotation}`}</span>
+      <WalletItemList moneyInfo={moneyState} onClickWon={handleClickWon} />
+      <TotalMoney moneyInfo={moneyState} />
     </>
   );
 }
